% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillgaze-package.R
\name{fill_gaze_gaps}
\alias{fill_gaze_gaps}
\title{Fill in gaps of missing eyetracking data}
\usage{
fill_gaze_gaps(data, ..., time_var = NULL, max_na_rows = NULL,
  max_duration = NULL, max_sd = NULL)
}
\arguments{
\item{data}{a dataframe of eyetracking data}

\item{...}{columns to interpolate selected using \link[dplyr:select_helpers]{dplyr selection
semantics}}

\item{time_var}{(optional) the name of a column in \code{data} containing the
timestamps of eyetracking samples. Defaults to using row numbers.}

\item{max_na_rows}{(optional) do not fill in gaps larger than \code{max_na_rows}
rows}

\item{max_duration}{(optional) do not fill in gaps longer than \code{max_na_rows}
in duration}

\item{max_sd}{(optional/experimental) do not fill in gaps where the change
from pre-missing to post-missing is more than \code{max_sd} * \emph{z}, where \emph{z} is
the standard deviation of frame-to-frame changes in values in that column.}
}
\value{
the dataframe with interpolated eyetracking data
}
\description{
Fill in gaps of missing eyetracking data
}
\details{
This function respects groupings created by \code{\link[dplyr:group_by]{dplyr::group_by()}}.
}
