% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillgaze-package.R
\name{find_gaze_gaps}
\alias{find_gaze_gaps}
\title{Find gaps in a column of eyetracking data}
\usage{
find_gaze_gaps(data, var, time_var = NULL)
}
\arguments{
\item{data}{a dataframe of eyetracking data}

\item{var}{the name of a column in \code{data} to check for gaps}

\item{time_var}{(optional) the name of a column in \code{data} containing the
timestamps of eyetracking samples. Defaults to using row numbers.}
}
\value{
a dataframe with one row per gap of eyetracking data.
}
\description{
Find gaps in a column of eyetracking data
}
\details{
This function respects groupings created by \code{\link[dplyr:group_by]{dplyr::group_by()}}.
}
