% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillgaze-package.R
\name{set_values_to_na}
\alias{set_values_to_na}
\title{Set values in dataframe columns to NA}
\usage{
set_values_to_na(data, ...)
}
\arguments{
\item{data}{a dataframe}

\item{...}{predicate functions that return true whenever a value should be
replaced with NAs. The functions should be named, so that the argument
\code{var1 = is.finite} would replace all the values in the column \code{var1} where
\code{is.finite()} returns \code{TRUE} with `NA`` values. These predicate functions
can be defined using the \link[rlang:as_function]{formula syntax for anonymous
functions}}
}
\value{
a modified copy of the dataframe
}
\description{
Set values in dataframe columns to NA
}
\examples{
is_zero <- function(x) x == 0
set_values_to_na(mtcars, cyl = ~ .x == 6, vs = is_zero)
}
